+++ package.json
@@ -6,6 +6,7 @@
     "@apollo/client": "^3.4.16",
     "@ethersproject/providers": "^5.4.1",
     "@ethersproject/strings": "^5.4.0",
+    "@react-hook/window-size": "^3.0.7",
     "@testing-library/jest-dom": "^5.11.4",
     "@testing-library/react": "^11.1.0",
     "@testing-library/user-event": "^12.1.10",
