+++ src/components/CollateralList/index.ts
@@ -1,0 +1,2 @@
+export { default as CollateralListContainer } from './CollateralListContainer';
+export { default as CollateralList } from './CollateralList';
