+++ src/components/CollateralList/mappingCollateralsData.ts
